requirements.txt
mediapipe
opencv-python
numpy
pyserial
groq
python-dotenv
